DBZ-5071 Correctly handle NULL values in incremental snapshots
This makes changes in the Informix connector for the corresponding
commit in the main Debezium project.  The main thing we have to do is
override the nullsSortLast function in InformixConnection to define how
NULLs sort in Informix.  We also have to update the test suite, since
the base test suite now includes additional tests for working with keys
that are nullable.